backend/requirements.txt
fastapi
uvicorn
networkx
google-genai
pydantic
pandas
